+++ include/csr.h
@@ -6,33 +6,33 @@
 
 /**
     * @brief Struct to represent a Compressed Sparse Row matrix
-    * @param {unsigned long} rows Number of rows in the matrix
-    * @param {unsigned long} cols Number of columns in the matrix
-    * @param {unsigned long} nnz Number of non-zero elements in the matrix
-    * @param {unsigned long[]} row_offsets Array of size rows + 1 containing the offsets of the rows
-    * @param {unsigned long[]} col_indices Array of size nnz containing the column indices of the non-zero elements
+    * @param {int} rows Number of rows in the matrix
+    * @param {int} cols Number of columns in the matrix
+    * @param {int} nnz Number of non-zero elements in the matrix
+    * @param {int[]} row_offsets Array of size rows + 1 containing the offsets of the rows
+    * @param {int[]} col_indices Array of size nnz containing the column indices of the non-zero elements
     * @param {float[]} values Array of size nnz containing the values of the non-zero elements
  */
 struct csr_matrix {
-    size_t rows;
-    size_t cols;
-    size_t nnz;
-    size_t* row_offsets;
-    size_t* col_indices;
-    float* values;
+    int rows;
+    int cols;
+    int nnz;
+    int* row_offsets; // Lenght: csr->rows + 1
+    int* col_indices; // Lenght: csr->nnz
+    float* values;       // Lenght: csr->nnz
 };
 
 /**
     * @brief Builds a new CSR matrix from the given data
-    * @param {unsigned long} rows Number of rows in the matrix
-    * @param {unsigned long} cols Number of columns in the matrix
-    * @param {unsigned long} nnz Number of non-zero elements in the matrix
-    * @param {unsigned long[]} row_offsets Array of size rows + 1 containing the offsets of the rows
-    * @param {unsigned long[]} col_indices Array of size nnz containing the column indices of the non-zero elements
+    * @param {int} rows Number of rows in the matrix
+    * @param {int} cols Number of columns in the matrix
+    * @param {int} nnz Number of non-zero elements in the matrix
+    * @param {int[]} row_offsets Array of size rows + 1 containing the offsets of the rows
+    * @param {int[]} col_indices Array of size nnz containing the column indices of the non-zero elements
     * @param {float[]} values Array of size nnz containing the values of the non-zero elements
     * @returns {csr_matrix *} Pointer to the CSR matrix
  */
-csr_matrix* new_csr_matrix(size_t rows, size_t cols, size_t nnz, size_t *row_offsets, size_t *col_indices, float *values);
+csr_matrix* new_csr_matrix(int rows, int cols, int nnz, int *row_offsets, int *col_indices, float *values);
 /**
     * @brief Builds a new CSR matrix with empty arrays
     * @param {unsigned long} rows Number of rows in the matrix
@@ -40,7 +40,7 @@ csr_matrix* new_csr_matrix(size_t rows, size_t cols, size_t nnz, size_t *row_off
     * @param {unsigned long} nnz Number of non-zero elements in the matrix
     * @returns {csr_matrix *} Pointer to the CSR matrix
  */
-csr_matrix* new_csr_matrix(size_t rows, size_t cols, size_t nnz);
+csr_matrix* new_csr_matrix(int rows, int cols, int nnz);
 /**
     * @brief Load a CSR matrix from a file
     * @param {const char *} filename Path to the file
